requirements.txt
alembic==0.9.3
aniso8601==1.2.1
click==6.7
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.1
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.11
Unidecode==0.4.21
Werkzeug==0.12.2
